Cosmetic changes and test improvement

+++ tests/conftest.py
@@ -26,8 +26,11 @@ class Solution:
     def __init__(self, directory):
         self.directory = directory
 
-    def run_in(self, cmd, path=""):
-        result =  subprocess.run(f"cd {self.directory}/{path} && {cmd}", shell=True, check=True)
+    def run_in(self, cmd, path="", capture=False):
+        if capture:
+            result =  subprocess.run(f"cd {self.directory}/{path} && {cmd}", shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        else:
+            result =  subprocess.run(f"cd {self.directory}/{path} && {cmd}", shell=True, check=True)
         if result.returncode == 0:
             return result.stdout
         else:
@@ -42,18 +45,26 @@ class Solution:
 
 @pytest.fixture(scope="class")
 def solution(request, tmp_path_factory):
-    level = request.node.get_closest_marker("level").args[0]
-    for mark in request.node.iter_markers():
-        # We iterate because we want the mark on the subclass
-        if mark.name == "level":
-            level = mark.args[0]
-    dir = tmp_path_factory.mktemp("base") / "test-solution"
-    os.mkdir(dir.resolve())
+    existing = os.getenv('SOLUTION')
+    if existing is None:
+        level = request.node.get_closest_marker("level").args[0]
+        for mark in request.node.iter_markers():
+            # We iterate because we want the mark on the subclass
+            if mark.name == "level":
+                level = mark.args[0]
+        dir = tmp_path_factory.mktemp("base") / "test-solution"
+        dir = dir.resolve()
+        os.mkdir(dir)
 
-    solution = Solution(dir.resolve())
-    solution.run_in(f"az-ai-generator-test '{dir.resolve()}' {level}")
-    yield solution
-    shutil.rmtree(dir)
+        solution = Solution(dir)
+        solution.run_in(f"az-ai-generator-test '{dir}' {level}")
+
+        yield solution
+
+        shutil.rmtree(dir)
+    else:
+      yield Solution(existing)
+
 
 
 @pytest.fixture(scope="class")
@@ -73,33 +84,40 @@ def depend_on_fast(request):
 
 @pytest.fixture(scope="class")
 def azd_env(request, solution):
+    if os.getenv('SOLUTION') is None:
+        env_name = solution.run_in("azd env get-value AZURE_ENV_NAME", capture=True).strip().decode('utf-8')
 
-    instance_count = os.getenv('AZURE_INSTANCE_COUNT', '001')
-    env_name = f"pyt-az-ai-{instance_count}"
-    location = os.getenv('AZURE_LOCATION', 'francecentral')
-    resource_group_name = f"rg-{env_name}"
-    app_name = f"{env_name}-app"
+        instance_count = os.getenv('AZURE_INSTANCE_COUNT', '001')
+        env_name = f"pyt-az-ai-{instance_count}"
+        location = os.getenv('AZURE_LOCATION', 'francecentral')
+        resource_group_name = f"rg-{env_name}"
+        app_name = f"{env_name}-app"
 
-    # Create the test environment
-    solution.run_in(f"azd env new {env_name}")
+        # Create the test environment
+        solution.run_in(f"azd env new {env_name}")
 
-    # Use the default subscription from Azure CLI
-    solution.run_in('azd env set AZURE_SUBSCRIPTION_ID "$(az account show --query id -o tsv)"')
-    solution.run_in(f"azd env set AZURE_RESOURCE_GROUP {resource_group_name}")
-    solution.run_in(f"azd env set AZURE_LOCATION {location}")
+        # Use the default subscription from Azure CLI
+        solution.run_in('azd env set AZURE_SUBSCRIPTION_ID "$(az account show --query id -o tsv)"')
+        solution.run_in(f"azd env set AZURE_RESOURCE_GROUP {resource_group_name}")
+        solution.run_in(f"azd env set AZURE_LOCATION {location}")
 
-    # Cleanup the previous run in case it's still here from a previous run
-    solution.run_in((
-        f'app_id="$(az ad app list --display-name {app_name} --query "[0].appId" -o tsv)" && '
-        f'[ -n "$app_id" ] && az ad app delete --id "$app_id" || true'
-    ))
-    solution.run_in((
-        f'(test "$(az group exists --name {resource_group_name})" = "true" && '
-        f'azd down --purge --force --no-prompt) || true'
-    ))
+        # Cleanup the previous run in case it's still here from a previous run
+        solution.run_in((
+            f'app_id="$(az ad app list --display-name {app_name} --query "[0].appId" -o tsv)" && '
+            f'[ -n "$app_id" ] && az ad app delete --id "$app_id" || true'
+        ))
+        solution.run_in((
+            f'(test "$(az group exists --name {resource_group_name} -o tsv)" = "true" && '
+            f'azd down --purge --force --no-prompt) || true'
+        ))
 
-    solution.run_in(f"az group create --location {location} --name {resource_group_name}")
+        solution.run_in(f"az group create --location {location} --name {resource_group_name}")
+        solution.run_in(f"azd env set AZURE_RESOURCE_GROUP {resource_group_name}")
 
-    yield env_name
+        yield env_name
 
-    solution.run_in("azd down --purge --force --no-prompt || true")
+        solution.run_in("azd down --purge --force --no-prompt || true")
+    else:
+        env_name = solution.run_in("azd env get-value AZURE_ENV_NAME", capture=True).strip().decode('utf-8')
+
+        yield env_name
